feat: move predicate handling to the utility files
