+++ filter.js
@@ -1,3 +1,5 @@
+import { predicateHandler } from './utils/predicateHandler';
+
 // filter _.filter(collection, [predicate=_.identity])
 // Iterates over elements of collection, returning an array of all elements predicate returns truthy for. The predicate is invoked with three arguments: (value, index|key, collection).
 // Note: This method returns a new array.
@@ -26,46 +28,17 @@ export function filter(array, predicate = item => item) {
   }
 
   const result = [];
-  let matchesFunc;
-  switch (typeof predicate) {
-    case 'function': {
-      matchesFunc = item => predicate(item);
-    }
-    break;
-    case 'object': {
-      if (Array.isArray(predicate)) { // `_.matchesProperty` iteratee shorthand.
-        if (predicate.length > 2) {
-          return [];
-        }
-        matchesFunc = (item => item[predicate[0]] === predicate[1]);
-      } else { // `_.matches` iteratee shorthand.
-        matchesFunc = item => {
-          const predicateKeys = Object.keys(predicate);
-          let matches = true;
-          for (let j = 0; j < predicateKeys.length; j += 1) {
-            if (item[predicateKeys[j]] !== predicate[predicateKeys[j]]) {
-              matches = false;
-              break;
-            }
-          }
-          return matches;
-        }
+  try {
+    const matchesFunc = predicateHandler(predicate);
+
+    for (let i = 0; i < array.length; i += 1) {
+      if (matchesFunc(array[i])) {
+        result.push(array[i]);
       }
     }
-    break;
-    case 'string': { // `_.property` iteratee shorthand
-      matchesFunc = item => item[predicate];
-    }
-    break;
-    default:
-      return [];
   }
-
-  for (let i = 0; i < array.length; i += 1) {
-    const matches = matchesFunc(array[i]);
-    if (matchesFunc(array[i])) {
-      result.push(array[i]);
-    }
+  catch {
+    return [];
   }
 
   return result;
